+++ src/core/lang/verilog/symbols/module.rs
@@ -27,9 +27,10 @@ use crate::core::lang::{
     },
     verilog::{
         error::VerilogError,
-        interface::{self, ParamList, PortList},
+        interface::{self, DataType, ParamList, PortList},
         token::{identifier::Identifier, operator::Operator},
     },
+    vhdl::token::VhdlTokenizer,
 };
 use serde_derive::Serialize;
 use std::iter::Peekable;
@@ -269,11 +270,254 @@ impl Module {
     }
 }
 
+use crate::core::lang::lexer::Tokenize;
+use crate::core::lang::verilog::interface::tokens_to_string;
 use crate::core::lang::vhdl::error::VhdlError;
 use crate::core::lang::vhdl::symbols::entity::Entity;
+use crate::core::lang::vhdl::token::delimiter::Delimiter as VhDelimiter;
+use crate::core::lang::vhdl::token::identifier::Identifier as VhIdentifier;
+use crate::core::lang::vhdl::token::keyword::Keyword as VhKeyword;
+use crate::core::lang::vhdl::token::VhdlToken as Vvt;
+use crate::core::lang::vhdl::token::VhdlToken;
 
 impl Module {
+    /// Builds an [Entity] from the structured HDL data.
     pub fn to_vhdl_entity(&self) -> Result<Entity, VhdlError> {
-        todo!()
+        // start with blank slate of VHDL tokens to build from
+        let mut tokens: Vec<Token<Vvt>> = Vec::new();
+
+        // assemble entity name
+        let name = match &self.name {
+            Identifier::Basic(s) => VhIdentifier::Basic(s.clone()),
+            Identifier::Escaped(s) => VhIdentifier::Extended(s.clone()),
+            Identifier::Directive(s) => VhIdentifier::Extended(s.clone()),
+            Identifier::System(s) => VhIdentifier::Extended(s.clone()),
+        };
+        tokens.push(Self::vh(Vvt::Identifier(name)));
+
+        // add IS keyword
+        tokens.push(Self::vh(Vvt::Keyword(VhKeyword::Is)));
+
+        // add generics
+        if self.parameters.len() > 0 {
+            // add the correct begining syntax
+            tokens.push(Self::vh(Vvt::Keyword(VhKeyword::Generic)));
+            tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::ParenL)));
+
+            // add each parameter
+            self.parameters.iter().for_each(|p| {
+                // add the port identifier
+                let name = match &p.get_name() {
+                    Identifier::Basic(s) => VhIdentifier::Basic(s.clone()),
+                    Identifier::Escaped(s) => VhIdentifier::Extended(s.clone()),
+                    Identifier::Directive(s) => VhIdentifier::Extended(s.clone()),
+                    Identifier::System(s) => VhIdentifier::Extended(s.clone()),
+                };
+                tokens.push(Self::vh(Vvt::Identifier(name)));
+                // add the ':'
+                tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::Colon)));
+
+                // add the datatype
+                tokens.push(Self::vh(Self::convert_datatype_to_vh(p.get_datatype())));
+
+                // any ranges for that dataype?
+                if let Some(ranges) = p.get_datatype().get_ranges() {
+                    ranges.into_iter().for_each(|r| {
+                        tokens.append(&mut Self::convert_range_to_vh(r.0, r.1));
+                    });
+                }
+
+                // add the default value (if exists)
+                if let Some(expr) = p.get_default().as_static_expr() {
+                    tokens.append(&mut Self::convert_default_to_vh(expr));
+                }
+
+                tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::Terminator)));
+            });
+
+            // remove the last trailing ';'
+            tokens.pop();
+
+            // add the correct closing syntax
+            tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::ParenR)));
+            tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::Terminator)));
+        }
+        // add ports
+        if self.ports.len() > 0 {
+            // add the correct begining syntax
+            tokens.push(Self::vh(Vvt::Keyword(VhKeyword::Port)));
+            tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::ParenL)));
+
+            // add each port
+            self.ports.iter().for_each(|p| {
+                // add the port identifier
+                let name = match &p.get_name() {
+                    Identifier::Basic(s) => VhIdentifier::Basic(s.clone()),
+                    Identifier::Escaped(s) => VhIdentifier::Extended(s.clone()),
+                    Identifier::Directive(s) => VhIdentifier::Extended(s.clone()),
+                    Identifier::System(s) => VhIdentifier::Extended(s.clone()),
+                };
+                tokens.push(Self::vh(Vvt::Identifier(name)));
+                // add the ':'
+                tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::Colon)));
+                // add the port direction
+                let dir = if let Some(dir) = p.get_mode() {
+                    match dir {
+                        Keyword::Input => VhKeyword::In,
+                        Keyword::Output => VhKeyword::Out,
+                        Keyword::Inout => VhKeyword::Inout,
+                        _ => panic!("unsupported port mode conversion"),
+                    }
+                } else {
+                    VhKeyword::In
+                };
+                tokens.push(Self::vh(Vvt::Keyword(dir)));
+
+                // add the datatype
+                tokens.push(Self::vh(Self::convert_datatype_to_vh(p.get_datatype())));
+
+                // any ranges for that dataype?
+                if let Some(ranges) = p.get_datatype().get_ranges() {
+                    ranges.into_iter().for_each(|r| {
+                        tokens.append(&mut Self::convert_range_to_vh(r.0, r.1));
+                    });
+                }
+
+                // add the default value (if exists)
+                if let Some(expr) = p.get_default().as_static_expr() {
+                    tokens.append(&mut Self::convert_default_to_vh(expr));
+                }
+
+                tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::Terminator)));
+            });
+
+            // remove the last trailing ';'
+            tokens.pop();
+
+            // add the correct closing syntax
+            tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::ParenR)));
+            tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::Terminator)));
+        }
+
+        // add proper closing before parsing
+        tokens.push(Self::vh(Vvt::Keyword(VhKeyword::End)));
+
+        // parse the assembled token list to create the module
+        let mut tokens = tokens.into_iter().peekable();
+        Entity::from_tokens(&mut tokens, Position::new())
+    }
+
+    /// Helps build SystemVerilog tokens from a VHDL context.
+    fn vh(token: Vvt) -> Token<Vvt> {
+        Token::new(token, Position::new())
+    }
+
+    /// Helps convert a SV token into its VHDL equivalent when dealing with
+    /// datatypes.
+    fn convert_datatype_to_vh(datatype: &DataType) -> Vvt {
+        let dtype = datatype.get_type();
+        let has_range = datatype.get_ranges().is_some();
+
+        let datatype = if let Some(dtype) = dtype {
+            match dtype {
+                SystemVerilogToken::Keyword(k) => match k {
+                    Keyword::Int | Keyword::Integer => {
+                        Vvt::Identifier(VhIdentifier::Basic("integer".to_string()))
+                    }
+                    Keyword::Byte => Vvt::Identifier(VhIdentifier::Basic("bit_vector".to_string())),
+                    Keyword::Logic => match has_range {
+                        true => {
+                            Vvt::Identifier(VhIdentifier::Basic("std_logic_vector".to_string()))
+                        }
+                        false => Vvt::Identifier(VhIdentifier::Basic("std_logic".to_string())),
+                    },
+                    Keyword::Bit => match has_range {
+                        true => Vvt::Identifier(VhIdentifier::Basic("bit_vector".to_string())),
+                        false => Vvt::Identifier(VhIdentifier::Basic("bit".to_string())),
+                    },
+                    Keyword::String => Vvt::Identifier(VhIdentifier::Basic("string".to_string())),
+                    _ => panic!("unsupported datatype keyword conversion to vhdl"),
+                },
+                SystemVerilogToken::Identifier(i) => match i {
+                    Identifier::Basic(s) => Vvt::Identifier(VhIdentifier::Basic(s.clone())),
+                    Identifier::Escaped(s) => Vvt::Identifier(VhIdentifier::Extended(s.clone())),
+                    Identifier::Directive(s) => Vvt::Identifier(VhIdentifier::Extended(s.clone())),
+                    Identifier::System(s) => Vvt::Identifier(VhIdentifier::Extended(s.clone())),
+                },
+                _ => panic!("unsupported datatype conversion to vhdl"),
+            }
+        } else {
+            match has_range {
+                true => Vvt::Identifier(VhIdentifier::Basic("std_logic_vector".to_string())),
+                false => Vvt::Identifier(VhIdentifier::Basic("std_logic".to_string())),
+            }
+        };
+        datatype
+    }
+
+    /// Helps convert a SV range (set of tokens) into a VHDL equivalent set of tokens
+    /// for specifying a range for a particular datatype.
+    fn convert_range_to_vh(
+        lhs: Vec<SystemVerilogToken>,
+        rhs: Vec<SystemVerilogToken>,
+    ) -> Vec<Token<VhdlToken>> {
+        let mut tokens = Vec::new();
+        // opening bracket
+        tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::ParenL)));
+
+        // left side of range
+        VhdlTokenizer::tokenize(&tokens_to_string(&lhs))
+            .into_iter()
+            .filter_map(|r| match r {
+                Ok(r) => Some(r),
+                Err(_) => None,
+            })
+            .filter(|r| r.as_type().is_eof() == false)
+            .for_each(|t| {
+                tokens.push(t);
+            });
+
+        // delimiter between range endpoints
+        tokens.push(Self::vh(Vvt::Keyword(VhKeyword::Downto)));
+
+        // right side of range
+        VhdlTokenizer::tokenize(&tokens_to_string(&rhs))
+            .into_iter()
+            .filter_map(|r| match r {
+                Ok(r) => Some(r),
+                Err(_) => None,
+            })
+            .filter(|r| r.as_type().is_eof() == false)
+            .for_each(|t| {
+                tokens.push(t);
+            });
+
+        // closing bracket
+        tokens.push(Self::vh(Vvt::Delimiter(VhDelimiter::ParenR)));
+
+        tokens
+    }
+
+    /// Helps convert a SV default value into a VHDL equivalent set of tokens.
+    fn convert_default_to_vh(expr: &Vec<SystemVerilogToken>) -> Vec<Token<VhdlToken>> {
+        let mut tokens = Vec::new();
+
+        VhdlTokenizer::tokenize(&tokens_to_string(&expr))
+            .into_iter()
+            .filter_map(|r| match r {
+                Ok(r) => Some(r),
+                Err(_) => None,
+            })
+            .filter(|r| r.as_type().is_eof() == false)
+            .for_each(|t| {
+                tokens.push(t);
+            });
+
+        // only introduce the '=' token if we successfully transfered the VHDL to SV
+        if tokens.len() > 0 {
+            tokens.insert(0, Self::vh(Vvt::Delimiter(VhDelimiter::VarAssign)));
+        }
+
+        tokens
     }
 }