+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 ## 0.23.3 - unreleased
 
 ### Features
+- adds support for mixed language Verilog, SystemVerilog to VHDL instantiations with `get` subcommand and the `--language vhdl` option
+- adds support for mixed language VHDL to SystemVerilog instantiations with `get` subcommand and the `--language sv` option
 - adds support for recognizing the modules instantiated within a SystemVerilog `bind` construct
 - writes all known environment variables set by Orbit to the target's `.env` file during the planning phase for a build process (which then gets loaded before performing the execution phase)
 - adds `ORBIT_TOP_JSON`, `ORBIT_TB_JSON`, and `ORBIT_DUT_JSON` environment variables to be set before a target's exeuction phase